Move blocked callback regression to adversarial tests
